build: 抽离 rollup 公共配置
