+++ packages/effects/rollup.config.js
@@ -1,44 +1,9 @@
-import commonjs from '@rollup/plugin-commonjs';
-import resolve from '@rollup/plugin-node-resolve';
-import replace from '@rollup/plugin-replace';
-import { swc, defineRollupSwcOption, minify } from 'rollup-plugin-swc3';
-import glslInner from '../../scripts/rollup-plugin-glsl-inner';
+import { getBanner, getPlugins } from '../../scripts/rollup-config-helper';
 import appxConfig from './rollup.appx.config';
 
 const pkg = require('./package.json');
-const banner = `/*!
- * Name: ${pkg.name}
- * Description: ${pkg.description}
- * Author: ${pkg.author}
- * Contributors: ${pkg.contributors.map(c => c.name).join(',')}
- * Version: v${pkg.version}
- */
-`;
-
-const defines = {
-  __VERSION__: JSON.stringify(pkg.version),
-  __DEBUG__: false,
-};
-const plugins = [
-  replace({
-    preventAssignment: true,
-    values: defines,
-  }),
-  glslInner(),
-  swc(
-    defineRollupSwcOption({
-      exclude: [],
-      jsc: {
-        loose: true,
-        externalHelpers: true,
-        target: 'es5',
-      },
-      sourceMaps: true,
-    }),
-  ),
-  resolve(),
-  commonjs(),
-];
+const banner = getBanner(pkg);
+const plugins = getPlugins(pkg);
 
 export default (commandLineArgs) => {
   return [
@@ -65,9 +30,7 @@ export default (commandLineArgs) => {
         banner,
         sourcemap: true,
       },
-      plugins: plugins.concat(
-        minify({ sourceMap: true })
-      ),
+      plugins: getPlugins(pkg, true),
     },
     ...appxConfig.map(config => ({ ...config, plugins: plugins.concat(config.plugins) }))
   ];
