+++ plugin-packages/alipay-downgrade/rollup.config.js
@@ -1,47 +1,12 @@
-import commonjs from '@rollup/plugin-commonjs';
-import resolve from '@rollup/plugin-node-resolve';
-import replace from '@rollup/plugin-replace';
-import { swc, defineRollupSwcOption, minify } from 'rollup-plugin-swc3';
-import glslInner from '../../scripts/rollup-plugin-glsl-inner';
+import { getBanner, getPlugins } from '../../scripts/rollup-config-helper';
 
 const pkg = require('./package.json');
-const banner = `/*!
- * Name: ${pkg.name}
- * Description: ${pkg.description}
- * Author: ${pkg.author}
- * Contributors: ${pkg.contributors.map(c => c.name).join(',')}
- * Version: v${pkg.version}
- */
-`;
-
-const defines = {
-  __VERSION__: JSON.stringify(pkg.version),
-  __DEBUG__: false,
-};
+const banner = getBanner(pkg);
+const plugins = getPlugins(pkg);
 const globals = {
   '@galacean/effects': 'ge',
 };
 const external = Object.keys(globals);
-const plugins = [
-  replace({
-    preventAssignment: true,
-    values: defines,
-  }),
-  glslInner(),
-  swc(
-    defineRollupSwcOption({
-      exclude: [],
-      jsc: {
-        loose: true,
-        externalHelpers: true,
-        target: 'es5',
-      },
-      sourceMaps: true,
-    }),
-  ),
-  resolve(),
-  commonjs(),
-];
 
 export default (commandLineArgs) => {
   return [{
@@ -72,8 +37,6 @@ export default (commandLineArgs) => {
       sourcemap: true,
     },
     external,
-    plugins: plugins.concat(
-      minify({ sourceMap: true })
-    ),
+    plugins: getPlugins(pkg, true),
   }];
 };
